refactor: RestaurantFilters 네임스페이스 추가
